Pushing local changes
The reason for this is because A) I added AOS to the ace3ds+ page, not really recommended but it's an option. And B) The files for the timebomb r4isdhc.com kernel have moved off of the main page for the flashcard archive, instead they are now in the archive branch on the github!